% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certeapi-package}
\alias{certeapi}
\alias{certeapi-package}
\title{certeapi: A Certe R Package for Providing an Application Programming Interface}
\description{
A Certe R Package for providing an application programming interface (API) using 'plumber'. This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certeapi}
  \item \url{https://github.com/certe-medical-epidemiology/certeapi}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
