% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{endpoints}
\alias{endpoints}
\title{Overview of API Endpoints}
\description{
This page provides an overview of all available API endpoints and their arguments. On the API server, this information is available at \verb{\{host\}/__docs__/}, where \code{{host}} is the API location, such as \code{"localhost"}, an IP address, or DNS name.
}
\section{Overview of endpoints}{

\subsection{\verb{/esbl_prediction} (method \code{GET}):}{

Description: Retrieve an ESBL ETEST prediction and reliability for a vector of MIC values.

Arguments:
\itemize{
\item \code{AMC}  Amoxicillin/clavulanic acid, can be left blank
\item \code{AMP}  Ampicillin, can be left blank
\item \code{TZP}  Piperacillin/tazobactam, can be left blank
\item \code{CXM}  Cefuroxime, can be left blank
\item \code{FOX}  Cefoxitin, can be left blank
\item \code{CTX}  Cefotaxime, can be left blank
\item \code{CAZ}  Ceftazidime, can be left blank
\item \code{GEN}  Gentamicin, can be left blank
\item \code{TOB}  Tobramycin, can be left blank
\item \code{TMP}  Trimethoprim, can be left blank
\item \code{NIT}  Nitrofurantoin, can be left blank
\item \code{FOS}  Fosfomycin, can be left blank
\item \code{CIP}  Ciprofloxacin, can be left blank
\item \code{IPM}  Imipenem, can be left blank
\item \code{MEM}  Meropenem, can be left blank
\item \code{COL}  Colistin, can be left blank
}
}
}

