% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{startup}
\alias{startup}
\title{Start API}
\usage{
startup(port = read_secret("api.port"))
}
\arguments{
\item{port}{port number}
}
\description{
This function starts the API server.
}
\details{
For cron on UNIX (Linux/macOS), use the following command:

\verb{Rscript -e "certeapi::startup()" &>> /path/api.log &}

Use \code{Rscript --vanilla -e} to \emph{not} load any user settings such as \code{.Rprofile}. The trailing \code{&} will make the script run in the background.
}
