% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{endpoints}
\alias{endpoints}
\title{Overview of API Endpoints}
\source{
This page was automatically generated from the actual API R file, currently installed here: \verb{/home/uscloud/certeapi/inst/api.R}.
}
\description{
This page provides an overview of all available API endpoints and their arguments.
}
\section{Overview of endpoints}{

\subsection{\verb{/echo} (method \code{GET}):}{

Description: Echo back the input

Arguments:
\itemize{
\item \code{msg}  The message to echo
}
}

\subsection{\verb{/plot} (method \code{GET}):}{

Description: Plot a histogram

Arguments:

Serializer: \code{png}
}

\subsection{\verb{/mean} (method \code{POST}):}{

Description: Return the mean of two numbers

Arguments:
\itemize{
\item \code{a}  The first number to add
\item \code{b}  The second number to add
}
}

\subsection{\verb{/sum} (method \code{GET}):}{

Description: Test sum 2

Arguments:
\itemize{
\item \code{a}  getal 1
\item \code{b}  getal 2
}
}

\subsection{\verb{/esbl_prediction} (method \code{GET}):}{

Description: Retrieve an ESBL ETEST prediction and reliability for a vector of MIC values.

Arguments:
\itemize{
\item \code{mo}  microorganism, a code or a name
}
}
}

